Move BreakPlaceholderInWorkHistoryComponent under candidate namespace
