+++ app/components/candidate_interface/break_placeholder_in_work_history_component.html.erb
