+++ spec/components/candidate_interface/break_placeholder_in_work_history_component_spec.rb
@@ -1,6 +1,6 @@
 require 'rails_helper'
 
-RSpec.describe BreakPlaceholderInWorkHistoryComponent do
+RSpec.describe CandidateInterface::BreakPlaceholderInWorkHistoryComponent do
   let(:work_break) { double }
 
   before do
@@ -10,19 +10,19 @@ RSpec.describe BreakPlaceholderInWorkHistoryComponent do
   end
 
   it 'renders the component with a link to explain break' do
-    result = render_inline(BreakPlaceholderInWorkHistoryComponent.new(work_break: work_break))
+    result = render_inline(CandidateInterface::BreakPlaceholderInWorkHistoryComponent.new(work_break: work_break))
 
     expect(result.text).to include('Explain break between January 2020 and May 2020')
   end
 
   it 'renders the component with a link to add another job' do
-    result = render_inline(BreakPlaceholderInWorkHistoryComponent.new(work_break: work_break))
+    result = render_inline(CandidateInterface::BreakPlaceholderInWorkHistoryComponent.new(work_break: work_break))
 
     expect(result.text).to include('add another job between January 2020 and May 2020')
   end
 
   it 'renders the component with the break length' do
-    result = render_inline(BreakPlaceholderInWorkHistoryComponent.new(work_break: work_break))
+    result = render_inline(CandidateInterface::BreakPlaceholderInWorkHistoryComponent.new(work_break: work_break))
 
     expect(result.text).to include('You have a break in your work history in the last 5 years (3 months)')
   end
